Added description and first build and config instructions to README
A first proper set of configuration and build instructions was added. For now
these instructions are limited to generating the ISO image version of the
backup system. The documentation about deploying the backup system to a USB
drive is still very poor and needs attention in the future.

The configuration contains details on how to setup the keyboad layout. The
keyboard section is almost done. Only paths on the USB-drive instructions are
still missing. The actual configuration of the backup behaviour is still
missing. Only the paths where files are located has been added.