Changed build process to use make and optionally docker
In an attempt to further automate the build process, make is now used to build
Backup System. This further enables the possibility to call the same build
commands inside different contexts. This is done in the new option to build
Backup System using a docker image.

The option to build Backup System using a docker image, was introduced for two
main reasons:
 1. It enables building Backup System on non Debian based operating systems.
 2. The live-build tools are quite rough to the operating system on which they
    run. They install and delete numerous packages on the building system. Not
    everyone is happy with that, and can now run this messy process inside a
    docker container.